+++ CHANGELOG.md
@@ -1,8 +1,15 @@
-# Changelog
+  # Changelog
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## 2.9.0
+
+### Added
+
+- Significant rewrite of the web UI.
+- Added support for direct integration of editing SuperSettings settings from the web UI when `super_settings_editing` is enabled.
 
 ## 2.8.1